Refactor server main.cpp
